Deleted sample.log and appended test_utils_log.py
